--- a/docs/content/rules/sort-imports.mdx
+++ b/docs/content/rules/sort-imports.mdx
@@ -201,7 +201,7 @@ Specifies a maximum line length for sorting imports. When the line length exceed
 
 This option is only available when the type is set to `'line-length'`.
 
-### tsConfigRootDir
+### tsconfigRootDir
 
 <sub>default: `undefined`</sub>