--- a/test/read-closest-ts-config-by-path.test.ts
+++ b/test/read-closest-ts-config-by-path.test.ts
@@ -35,7 +35,7 @@ vi.mock('../utils/get-typescript-import', () => ({
 
 const testInput = {
   filePath: './repos/repo/packages/package/file.ts',
-  tsConfigRootDir: './repos/repo',
+  tsconfigRootDir: './repos/repo',
 }
 
 const tsConfigContent = {
@@ -133,14 +133,14 @@ describe('readClosestTsConfigByPath', () => {
         // This should call to fs.existsSync three times: c, b, a
         readClosestTsConfigByPath({
           filePath: './a/b/c/d.ts',
-          tsConfigRootDir: './a',
+          tsconfigRootDir: './a',
         })
 
         // This should call to fs.existsSync once: e
         // Then it should retrieve c from cache, pointing to a
         let actual = readClosestTsConfigByPath({
           filePath: './a/b/c/e/f.ts',
-          tsConfigRootDir: './a',
+          tsconfigRootDir: './a',
         })
 
         expect(actual).toEqual(tsConfigContent.raw.config.compilerOptions)
@@ -158,14 +158,14 @@ describe('readClosestTsConfigByPath', () => {
         // This should call to fs.existsSync 4 times: d, c, b, a
         readClosestTsConfigByPath({
           filePath: './a/b/c/d/e.ts',
-          tsConfigRootDir: './a',
+          tsconfigRootDir: './a',
         })
 
         // This should call to fs.existsSync 2: g, f
         // Then it should retrieve b from cache, pointing to a
         let actual = readClosestTsConfigByPath({
           filePath: './a/b/f/g/h.ts',
-          tsConfigRootDir: './a',
+          tsconfigRootDir: './a',
         })
 
         expect(actual).toEqual(tsConfigContent.raw.config.compilerOptions)
@@ -210,13 +210,13 @@ describe('readClosestTsConfigByPath', () => {
         )
       })
 
-      it('throws when searching passes the tsConfigRootDir', () => {
+      it('throws when searching passes the tsconfigRootDir', () => {
         mockExistsSync.mockReturnValue(false)
         mockReadConfigFileReturnValue()
         mockParseJsonConfigFileContentReturnValue()
 
         expect(() =>
-          readClosestTsConfigByPath({ ...testInput, tsConfigRootDir: '/' }),
+          readClosestTsConfigByPath({ ...testInput, tsconfigRootDir: '/' }),
         ).toThrowErrorMatchingInlineSnapshot(
           `[Error: Couldn't find any tsconfig.json relative to './repos/repo/packages/package/file.ts' within '/'.]`,
         )