Resolves #317

## Description

Adds a `tsconfigRootDir?: string` option (can be changed if needed). When passed by the user, each time a file is linted, the nearest `tsconfig.json` file will be searched, read, and used to determine if an import is internal or not.

Will try to dynamically load (and cache) `typescript` so this should not impact non-typescript users.

The `tsconfig.json` search and cache algorithm was taken from https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/typescript-estree/src/parseSettings/getProjectConfigFiles.ts 🙏. 

##  Process

- Linting is run for a given file.
- Check if `tsconfigRootDir` is entered. If not, sort as usual.
- Ensures that `typescript` is present. If not, sort as usual.
- If it is, search and read the closest `tsconfig.json` file and get the compiler options.
- If no file is found, sort as usual. If an error occurs (config file is invalid, for example), throw an error.
- For each import, use the compiler options to determine if that import can be marked as `internal` or `internal-type` (for the `groups` option).

## Assigning the `internal` / `internal-type` / `external` / `external-type` group to an import

Reminder: currently, `internal` and `internal-type` are **only** assigned to groups that match the `internalPattern` option. Non-external imports are not automatically matched with the `internal` / `internal-type` group.

I have chosen to keep this, and add the following:
If an import is considered 'internal' according to the user's `tsconfig`, that import will be matched with the `internal` / `internal-type` group as if it was matching `internalPattern`.

Everything below also applies for `type` imports with `internal-type` and `external-type`.

For a given import:

- If `typescript` is not present, use the standard process: the import is `external` if it starts with `.` or `/`.
- Else: If `typescript.isExternalModuleNameRelative(import)`, do not assign `internal` nor `external` at this stage.
- If `tsconfigRootDir` is not entered or if no `tsconfig.json` file was found, assign `external`.
- Get typescript information about the import using `typescript.resolveModuleName`.
- If the resolution fails, assign `external`.
- If the resolution succeeds, assign `external` if `resolution.resolvedModule.isExternalLibraryImport === true`, otherwise `internal`.

## Performance

Tests on a ~10k files codebase, 1 `tsconfig.json`.

#### Without caching `resolveModuleName` attempts

![image](https://github.com/user-attachments/assets/a4697d55-e4d6-4bf8-844e-46c8ae31ccda)

#### With caching `resolveModuleName` attempts

![image](https://github.com/user-attachments/assets/c85a83c2-a81e-48df-a278-3aa1a0636343)

#### Without `tsconfigRootDir` option (what we have today)

(Some other rules were disabled)

![image](https://github.com/user-attachments/assets/01dc332c-2d73-4399-8a1e-72adcac38e12)

#### Bottlenecks

Tests made while caching `resolveModuleName` attempts (~950ms)

- Without searching `tsconfig.json`: ~900ms
- Without `typescript.resolveModuleName`: ~620ms

#### Conclusion

Some significant performance hit despite caching.

- Searching for the appropriate `tsconfig.json`: ~+15%
- `typescript.resolveModuleName`: ~+75%
- Reading and parsing the config: ~300ms? 

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

- [x] New Feature