@hugop95 2024-11-04
I've run some performance tests, you can see the results in the PR description.

Long story short, there is going to be a performance hit:
- Searching for the appropriate `tsconfig.json` file is fast (<100ms in my test), but note that I tested it in a large project with a single file. Multiple `tsconfig.json` shouldn't affect the performance at this stage too much, though.
-  Reading/parsing `tsconfig.json`. I haven't tested that single stage too far, but I wouldn't be surprised if this takes 200+ms for large projects, or even more.
- `typescript.resolveModuleName`: the function used to know whether an import is external or not according to typescript. Even with cache, this takes a non-negligible amount of time, probably ~300ms in my tests.

Without the option activated, the total time for everything was about 350ms.