@hugop95 2024-11-04
I would actually love to be able to get feedback on whether this is going to be an issue or not.

@gavvvr Would you be okay for testing this in your project(s)? This will also be a great opportunity to make sure that the new algorithm works as expected in practice.