@gavvvr 2024-11-04
> @gavvvr Would you be okay for testing this in your project(s)? This will also be a great opportunity to make sure that the new algorithm works as expected in practice.

Hello @hugop95. I don't have much code to test this. There are only several absolute imports in [fd8e9a2ec](https://github.com/gavvvr/obsidian-imgur-plugin/tree/fd8e9a2ec4858606d9e240f4d4401514711e2ffc) snapshot of my open-source project. Right after this commit I switched to relative imports, because I wanted sorting/grouping by perfectionist plugin to work as expected :)

Those absolute imports appeared in my codebase only because I used vscode's hotkey to add missing imports. If you are still interested, that probably will be faster for you to test it (now you know where the codebase is). Unfortunately, I'm not much experienced with testing things, which are not yet on npm registry.