@hugop95 2024-11-05
@azat-io @OlivierZal  
I have finished my work on this PR, performance concerns are addressed at the bottom of the PR description.

Overall, I have mixed feelings on this:

The good points:
- Code specific to handling `typescript` import, searching for the `tsconfig.json` files, etc... are isolated from the rule, so not that much code added in the rule itself.
- Performance impact is **ok**. It does make the rule take a longer time, but it's still acceptable.   

The bad:
- It's still a lot of code for one option in one rule.
- I haven't been able to test this in a large repository with multiple configs (real use-case), so even if we have tests, they don't replicate complex use cases and we might still have bugs.
-  I don't know if it will be used: will it be worth maintaining in the end?

Still to discuss:
- Are we keeping `tsconfigRootDir ` as the option's name (https://github.com/azat-io/eslint-plugin-perfectionist/pull/345#discussion_r1827315159)? What about the default value?