@OlivierZal 2024-11-05
Current implementation looks good to me:
- `tsconfigRootDir` name is consistent with existing eslint options (we could eventually also authorize the value `true`, like this is sometimes the case with some eslint options – e.g. `projectService` –, which would read `languageOptions.parserOptions.tsconfigRootDir`)
- not enabled by default, so no impact on performances if the user doesn't explicitly fill it