@azat-io 2024-11-09
It works!

Perhaps we should also consider support for aliases from `paths` in `tsconfig.json` in the future. Thank you!